Done with the version 1.0 of the project.
